+++ extensions/src/extensions.ts
@@ -11,3 +11,4 @@ export * from './extensions/setPage';
 export * from './extensions/reinitializeOptions';
 export * from './extensions/setSyncData';
 export * from './extensions/invalidateBodyContent';
+export * from './extensions/showMetadataSelector';